requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
streamlit==1.40.0
requests
pillow
streamlit-drawable-canvas